useAuth moved to hooks
