+++ src/helpers/ProtectedRoute.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Navigate, Outlet} from 'react-router-dom';
 
-import { useAuth} from '../providers/AuthProvider';
+import { useAuth} from '../hooks/useAuth';
 
 export const ProtectedRoute = () => {
     const { user } = useAuth();
